requirements.txt
Flask
gunicorn
torch
torchvision
pennylane
Pillow